refactor utils, routes and add pino middleware logger. (#25)
* refactor utils, routes and add pino middleware logger.
* update readme
