+++ server/package.json
@@ -17,6 +17,7 @@
     "cookie-parser": "^1.4.3",
     "dotenv": "^5.0.1",
     "express": "^4.16.2",
+    "express-pino-logger": "3.0.2",
     "jwt-simple": "0.5.1",
     "moment": "2.22.2",
     "mongoose": "5.1.7",
